+++ R/rgho.R
@@ -7,6 +7,14 @@
 #'
 #' @docType package
 #' @name rgho
+#' @section Package Options:
+#'
+#'   \itemize{ \item \code{rgho.verbose} Should informations
+#'   about the request be diplayed? Defaults to
+#'   \code{FALSE}. \item \code{rgho.retry} How many times
+#'   the request should be retried? Defaults to 5. \item
+#'   \code{rgho.memotime} For how long should the results of
+#'   a request be memoised, in seconds? Defaults to 3600. }
 #'
 #' @importFrom httr parse_url
 #' @importFrom httr build_url
@@ -15,13 +23,9 @@
 #' @importFrom httr user_agent
 #'
 #' @importFrom memoise memoise
+#' @importFrom memoise timeout
 #'
 #' @importFrom curl ie_get_proxy_for_url
-#'
-#' @importFrom xml2 xml_find_all
-#' @importFrom xml2 xml_attr
-#' @importFrom xml2 xml_contents
-#' @importFrom xml2 xml_text
 #'
 #' @importFrom readr read_csv
 #'
@@ -30,16 +34,21 @@
 #' @importFrom magrittr "%<>%"
 #' @importFrom magrittr "extract2"
 #'
-#' @importFrom dplyr data_frame
+#' @importFrom tibble as_tibble
+#'
 #' @importFrom dplyr filter_
 #' @importFrom dplyr bind_rows
+#' @importFrom dplyr select_
 #'
 #' @importFrom tidyr spread_
 #'
 #' @importFrom lazyeval lazy_dots
 #'
 #' @importFrom stats runif
+#' @importFrom stats setNames
 #'
 #' @importFrom utils head
+#'
+#' @importFrom jsonlite fromJSON
 #'
 NULL