+++ tests/testthat/test-search.R
@@ -4,9 +4,10 @@ test_that("search_dimension & search_codes work", {
 
   res1 <- search_dimensions("region")
   expect_identical(
-    head(res1),
-    c("GBDREGION", "REGION", "SUBREGION", "UNMDGREGION", "UNREGION",
-      "WHOINCOMEREGION")
+    sort(res1),
+    c("DHSMICSGEOREGION", "GBDREGION", "MGHEREG", "OECDREGION", "REGION",
+      "SUBREGION", "UNMDGREGION", "UNREGION", "WHOINCOMEREGION", "WORLDBANKREGION"
+    )
   )
   expect_identical(
     length(res1), 10L
@@ -14,9 +15,9 @@ test_that("search_dimension & search_codes work", {
 
   res2 <- search_codes("neonatal", dimension = "GHO")
   expect_identical(
-    head(res2),
-    c("WHS3_56", "WHOSIS_000003", "WHS4_128", "WHS2_515", "CM_03",
-      "nmr")
+    sort(res2),
+    sort(c("CM_03", "nmr", "pnmr", "WHOSIS_000003", "WHS2_515", "WHS3_56",
+      "WHS4_128"))
   )
   expect_identical(
     length(res2), 7L
@@ -25,9 +26,9 @@ test_that("search_dimension & search_codes work", {
   result <- get_gho_codes(dimension = "REGION")
   res3 <- search_gho(result, "asia")
   expect_identical(
-    head(res3),
-    c("SEAR", "WHO_LMI_SEAR", "GBD_REG14_SEARB", "GBD_REG14_SEARD",
-      "WHO_HI_SEAR")
+    sort(res3),
+    c("GBD_REG14_SEARB", "GBD_REG14_SEARD", "SEAR", "WHO_HI_SEAR",
+      "WHO_LMI_SEAR")
   )
   expect_identical(
     length(res3),