+++ NEWS.md
@@ -1,3 +1,16 @@
+# rgho 0.2.1
+
+## Backend changes
+
+  * `verbose` and `retry` can now be set by `options()`.
+  * Memoise results for a limited time (1h by default).
+  * Technical details are explained in a vignette.
+  * Package switched from XML to json.
+  
+## Bugfixes
+
+  * Test do not fail when new codes are added.
+
 # rgho 0.2.0
 
   * Fix encoding problem with country names.